Added cache, logs and cleanup
