+++ testframework/src/main/java/net/neoforged/testframework/impl/test/AbstractTest.java
@@ -39,9 +39,9 @@ import net.neoforged.testframework.annotation.WithListener;
 import net.neoforged.testframework.gametest.EmptyTemplate;
 import net.neoforged.testframework.gametest.ExtendedGameTestHelper;
 import net.neoforged.testframework.gametest.GameTestData;
-import net.neoforged.testframework.gametest.GameTestHelperFactory;
 import net.neoforged.testframework.gametest.StructureTemplateBuilder;
 import net.neoforged.testframework.impl.EventListenerGroupImpl;
+import net.neoforged.testframework.impl.GameTestHelperFactory;
 import net.neoforged.testframework.impl.MutableTestFramework;
 import net.neoforged.testframework.impl.ReflectionUtils;
 import net.neoforged.testframework.impl.TestFrameworkImpl;
